requirements.txt
livekit-agents
python-dotenv
chromadb
sentence-transformers
